# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pauza API
  description: Backend HTTP API for the Pauza focus-timer application.
  version: 0.1.0
  contact:
    name: Pauza Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.pauza.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token issued by POST /api/v1/auth/verify or POST /api/v1/auth/refresh.
    AdminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token issued by POST /api/v1/admin/login.
    WebhookBearer:
      type: http
      scheme: bearer
      description: Static secret shared with RevenueCat for webhook authentication.
  headers:
    XRequestId:
      description: Request identifier echoed in every response.
      schema:
        type: string
    XRateLimitLimit:
      description: Maximum requests allowed in the current window when the limiter is enforcing a real budget.
      schema:
        type: integer
    XRateLimitRemaining:
      description: Remaining requests in the current window when the limiter is enforcing a real budget.
      schema:
        type: integer
    XRateLimitReset:
      description: UTC Unix timestamp when the current window resets, when the limiter is enforcing a real budget.
      schema:
        type: integer

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    RegisterDeviceRequest:
      type: object
      additionalProperties: false
      required: [fcm_token, platform]
      description: Unknown JSON fields are rejected.
      properties:
        fcm_token:
          type: string
          minLength: 1
        platform:
          type: string
          enum: [android, ios]
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request validation failed (422).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Unexpected server error.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/devices:
    post:
      operationId: registerDevice
      tags: [Social]
      summary: Register a push notification device
      description: Unknown JSON fields are rejected. Requires a user JWT and is rate limited by user ID using the `general-api` tier. Tokens that Firebase later reports as unregistered are automatically pruned from the server.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "200":
          description: Device registered.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/devices/unregister:
    post:
      operationId: unregisterDevice
      tags: [Social]
      summary: Unregister a push notification device
      description: Unknown JSON fields are rejected. Requires a user JWT and is rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnregisterDeviceRequest"
      responses:
        "200":
          description: Device unregistered.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Social — Friends
  # =========================================================================
  /api/v1/friends:
    get:
      operationId: listFriends
      tags: [Social]
      summary: List accepted friends
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier. Invalid `page` and `limit` values silently fall back to defaults instead of returning 422.
      security:
        - UserJWT: []
      parameters:
        - name: page
          in: query
          description: 1-based page number. Invalid or non-positive values fall back to 1.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Page size. Invalid, non-positive, or values above 100 fall back to 20.
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated friend list.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/request:
    post:
      operationId: requestFriend
      tags: [Social]
      summary: Send a friend request
      description: Unknown JSON fields are rejected. Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestRequest"
      responses:
        "201":
          description: Friend request created.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendMutationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/requests/incoming:
    get:
      operationId: listIncomingRequests
      tags: [Social]
      summary: List incoming friend requests
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: Pending incoming requests.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/requests/outgoing:
    get:
      operationId: listOutgoingRequests
      tags: [Social]
      summary: List outgoing friend requests
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: Pending outgoing requests.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/requests/{id}/accept:
    post:
      operationId: acceptFriend
      tags: [Social]
      summary: Accept a friend request
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friendship identifier. The handler only checks that the path segment is non-empty.
          schema:
            type: string
      responses:
        "200":
          description: Request accepted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendMutationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/requests/{id}/decline:
    post:
      operationId: declineFriend
      tags: [Social]
      summary: Decline a friend request
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friendship identifier. The handler only checks that the path segment is non-empty.
          schema:
            type: string
      responses:
        "200":
          description: Request declined.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/requests/{id}/cancel:
    post:
      operationId: cancelFriendRequest
      tags: [Social]
      summary: Cancel an outgoing friend request
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friendship identifier. The handler only checks that the path segment is non-empty.
          schema:
            type: string
      responses:
        "200":
          description: Request cancelled.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/{id}:
    delete:
      operationId: removeFriend
      tags: [Social]
      summary: Remove an accepted friend
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friendship identifier. The handler only checks that the path segment is non-empty.
          schema:
            type: string
      responses:
        "200":
          description: Friend removed.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/{id}/stats:
    get:
      operationId: friendStats
      tags: [Social]
      summary: Get a friend's focus statistics
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier. Invalid `days` values fall back to 7 instead of returning 422.
      security:
        - UserJWT: []
      parameters:
        - name: id
          in: path
          required: true
          description: Friendship identifier. The handler only checks that the path segment is non-empty.
          schema:
            type: string
        - name: days
          in: query
          description: Rolling window in days. Values outside 1..90 or non-integers fall back to 7.
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Friend statistics returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/friends/search:
    get:
      operationId: searchUsers
      tags: [Social]
      summary: Search users by username prefix
      description: Requires a user JWT and the `premium` entitlement. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: q
          in: query
          required: true
          description: Username prefix (min 3, max 50 characters).
          schema:
            type: string
            minLength: 3
            maxLength: 50
      responses:
        "200":
          description: Matching users returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchUsersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/SubscriptionRequired"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Social — Leaderboards
  # =========================================================================
  /api/v1/leaderboard/streaks:
    get:
      operationId: leaderboardStreaks
      tags: [Social]
      summary: Leaderboard ranked by current streak
      description: Ranks users by current streak. Only users with `leaderboard_visible=true` appear in `entries` with contiguous ranks. `my_rank` is computed among visible users plus the caller (so hidden users can see their own position). Requires a user JWT and is rate limited by user ID using the `general-api` tier. Invalid `page` and `limit` values silently fall back to defaults.
      security:
        - UserJWT: []
      parameters:
        - name: page
          in: query
          description: 1-based page number. Invalid or non-positive values fall back to 1.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Page size. Invalid, non-positive, or values above 100 fall back to 20.
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Streak leaderboard returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...